Jules wip 6113416514066173746

+++ app/models.py
@@ -21,8 +21,8 @@ class KRCreateRequest(BaseModel):
     kr_number: int
     title: str
     description: str # Detailed description of the KR itself
-    meta_prevista: float = Field(..., ge=0, le=100) # Percentage
-    meta_realizada: float = Field(default=0.0, ge=0, le=100) # Percentage
+    meta_prevista: int = Field(..., ge=0, le=100) # Percentage
+    meta_realizada: int = Field(default=0, ge=0, le=100) # Percentage
     team_label: str
     product_label: str
     responsaveis: List[str]
@@ -39,8 +39,8 @@ class KRDescriptionUpdateRequest(BaseModel):
 
 class KRUpdateRequest(BaseModel):
     description: Optional[str] = None
-    meta_prevista: Optional[float] = Field(default=None, ge=0, le=100) # Percentage
-    meta_realizada: Optional[float] = Field(default=None, ge=0, le=100) # Percentage
+    meta_prevista: Optional[int] = Field(default=None, ge=0, le=100) # Percentage
+    meta_realizada: Optional[int] = Field(default=None, ge=0, le=100) # Percentage
     responsaveis: Optional[List[str]] = None
 
 # --- Activity Models ---
@@ -49,8 +49,8 @@ class Activity(BaseModel):
     stakeholders: str
     deadline_planned: str # e.g., "Month/Year" or specific date string
     deadline_achieved: Optional[str] = None
-    progress_planned_percent: float = Field(..., ge=0, le=100)
-    progress_achieved_percent: float = Field(default=0.0, ge=0, le=100)
+    progress_planned_percent: int = Field(..., ge=0, le=100)
+    progress_achieved_percent: int = Field(default=0, ge=0, le=100)
 
 class ActivityCreateRequest(BaseModel):
     activities: List[Activity]
